chore: refactor components out into directories, use AuthPage
