+++ frontend/routes/auth/wallets/key-details/assets-list/asset-card.spec.tsx
