+++ frontend/routes/auth/wallets/key-details/assets-list/asset-card.tsx
@@ -1,8 +1,8 @@
-import { useAssetsStore } from '../../../../stores/assets-store'
+import { useAssetsStore } from '../../../../../stores/assets-store'
 import { addDecimalsFormatNumber, formatNumber, toBigNum } from '@vegaprotocol/utils'
-import { DataTable } from '../../../../components/data-table/data-table'
+import { DataTable } from '../../../../../components/data-table/data-table'
 import BigNumber from 'bignumber.js'
-import { CollapsibleCard } from '../../../../components/collapsible-card'
+import { CollapsibleCard } from '../../../../../components/collapsible-card'
 import { MarketLozenges } from './markets-lozenges'
 import { vegaAccount, vegaAccountType } from '@vegaprotocol/rest-clients/dist/trading-data'
 import get from 'lodash/get'
