+++ frontend/routes/auth/wallets/key-details/assets-list/index.tsx
@@ -1,0 +1,1 @@
+export * from './assets-list'
