+++ frontend/routes/auth/wallets/key-details/assets-list/market-lozenges.spec.tsx
