+++ frontend/routes/auth/wallets/key-details/assets-list/use-accounts.spec.tsx
@@ -2,7 +2,7 @@ import { renderHook } from '@testing-library/react'
 import { useAccounts } from './use-accounts'
 import { useAccountsStore } from './accounts-store'
 import { vegaAccountType } from '@vegaprotocol/rest-clients/dist/trading-data'
-import { mockStore } from '../../../../test-helpers/mock-store'
+import { mockStore } from '../../../../../test-helpers/mock-store'
 const MOCK_KEY = '1'.repeat(64)
 const ASSET_ID = '2'.repeat(64)
 const MARKET_ID = '3'.repeat(64)
@@ -31,7 +31,7 @@ jest.mock('./accounts-store', () => ({
 }))
 
 describe('UseAccounts', () => {
-  it('fetches the party, starts polling and returns the key and assets', () => {
+  it('fetches the party, starts polling and returns the assets', () => {
     const fetchAccounts = jest.fn()
     const startPoll = jest.fn()
     const stopPoll = jest.fn()
@@ -56,13 +56,7 @@ describe('UseAccounts', () => {
     })
 
     const view = renderHook(() => useAccounts(MOCK_KEY))
-    const { key, accountsByAsset } = view.result.current
-    expect(key).toStrictEqual({
-      publicKey: MOCK_KEY,
-      name: 'test',
-      index: 0,
-      metadata: []
-    })
+    const { accountsByAsset } = view.result.current
     expect(accountsByAsset).toStrictEqual({
       [ASSET_ID]: [
         {
