+++ frontend/routes/auth/wallets/key-details/key-details-page.tsx
@@ -1,23 +1,26 @@
 import { NavLink } from 'react-router-dom'
 import { ArrowLeft } from '../../../../components/icons/arrow-left'
 import { VegaKey } from '../../../../components/keys/vega-key'
-import { VegaSection } from '../../../../components/vega-section'
 import { useAssetsStore } from '../../../../stores/assets-store'
 import { useWalletStore } from '../../../../stores/wallets'
 import { FULL_ROUTES } from '../../../route-names'
-import { AssetCard } from './asset-card'
 import { KeySelector } from './key-selector'
-import { useAccounts } from './use-accounts'
-import { SubHeader } from '../../../../components/sub-header'
-import { ExportPrivateKeysDialog } from './export-private-keys-dialog'
+import { ExportPrivateKeysDialog } from './export-private-key-dialog'
+import { AssetsList } from './assets-list'
+import { AuthPage } from '../../../../components/auth-page'
+import { VegaSection } from '../../../../components/vega-section'
 
 export const locators = {
   keyDetailsDescription: 'key-details-description',
-  keyDetailsBack: 'key-details-back'
+  keyDetailsBack: 'key-details-back',
+  keyDetailsPage: 'key-details-page'
 }
 
 export const KeyDetailsPage = ({ id }: { id: string }) => {
-  const { accountsByAsset, key } = useAccounts(id)
+  const { getKeyById } = useWalletStore((state) => ({
+    getKeyById: state.getKeyById
+  }))
+  const key = getKeyById(id)
   const { loading } = useAssetsStore((state) => ({
     loading: state.loading
   }))
@@ -28,8 +31,9 @@ export const KeyDetailsPage = ({ id }: { id: string }) => {
   if (!key) throw new Error(`Key with id ${id} not found`)
 
   return (
-    <div>
-      <section>
+    <AuthPage
+      dataTestId={locators.keyDetailsPage}
+      title={
         <div className="mb-6 flex items-center text-white">
           <NavLink
             data-testid={locators.keyDetailsBack}
@@ -40,18 +44,11 @@ export const KeyDetailsPage = ({ id }: { id: string }) => {
           </NavLink>
           <KeySelector currentKey={key} />
         </div>
-        <VegaKey publicKey={key.publicKey} />
-      </section>
-      <VegaSection>
-        <SubHeader content="Balances" />
-        <p data-testid={locators.keyDetailsDescription} className="text-vega-dark-400 my-3">
-          Recent balance changes caused by your open positions may not be reflected below
-        </p>
-        {Object.entries(accountsByAsset).map(([assetId, val]) => (
-          <AssetCard key={assetId} accounts={val} assetId={assetId} />
-        ))}
-      </VegaSection>
+      }
+    >
+      <VegaKey publicKey={key.publicKey} />
+      <AssetsList id={id} />
       <ExportPrivateKeysDialog />
-    </div>
+    </AuthPage>
   )
 }
