+++ frontend/routes/auth/wallets/key-details/key-selector/index.tsx
@@ -1,0 +1,1 @@
+export * from './key-selector'
