+++ frontend/routes/auth/wallets/key-details/key-selector/key-selector.spec.tsx
