+++ frontend/routes/auth/wallets/key-details/key-selector/key-selector.tsx
@@ -1,8 +1,8 @@
 import { DropdownMenu, DropdownMenuContent, DropdownMenuTrigger } from '@vegaprotocol/ui-toolkit'
-import { IconChevronDown } from '../../../../components/icons/chevron-down'
-import { Key, useWalletStore } from '../../../../stores/wallets'
+import { IconChevronDown } from '../../../../../components/icons/chevron-down'
+import { Key, useWalletStore } from '../../../../../stores/wallets'
 import { useState } from 'react'
-import { KeyList } from '../../../../components/key-list'
+import { KeyList } from '../../../../../components/key-list'
 
 export const locators = {
   keySelectedCurrentKey: (keyName: string) => `${keyName}-selected-current-key`,
